+++ CHANGELOG.md
@@ -56,6 +56,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Widget_item : Product designation overflow
 - Share_one and Mainpage : Wishlist share process indicator visible on short process
 - Card_page : Big loyalty card render problem
+- Loyalty Cards Page : User can now scan different barcode types
 
 ### Security